+++ packages/terra-menu/tests/nightwatch/menu/SelectableAndUnselectableItemsMenu.jsx
@@ -12,6 +12,10 @@ class SelectableMenu extends React.Component {
     this.state = { open: true, isSelected: false };
   }
 
+  componentDidMount() {
+    this.forceUpdate();
+  }
+
   setButtonNode(node) {
     this.buttonNode = node;
   }