+++ packages/terra-menu/tests/nightwatch/menu/SmallMenu.jsx
@@ -11,6 +11,10 @@ class SmallMenu extends React.Component {
     this.state = { open: true };
   }
 
+  componentDidMount() {
+    this.forceUpdate();
+  }
+
   setButtonNode(node) {
     this.buttonNode = node;
   }