+++ packages/terra-popup/src/PopupWidths.js
@@ -1,0 +1,11 @@
+const PopupWidths = {
+  auto: -1,
+  160: 160,
+  240: 240,
+  320: 320,
+  640: 640,
+  960: 960,
+  1280: 1280,
+  1760: 1760,
+};
+export default PopupWidths;