+++ packages/terra-hookshot/CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 =========
+### Fixed
+* Mobile zoom on safari
+* Half px value causing blurriness
 
 Unreleased
 ----------