+++ packages/terra-hookshot/CHANGELOG.md
@@ -1,12 +1,16 @@
 Changelog
 =========
+
+
+Unreleased
+----------
+
+1.1.0 - (October 24, 2017)
+------------------
 ### Fixed
 * Mobile zoom on safari
 * Half px value causing blurriness
 * Disable Listener dependency on target
-
-Unreleased
-----------
 
 1.0.0 - (October 12, 2017)
 ------------------